Refactor: Extracted business logic into useInvites hook
