+++ client-interface/lib/hooks/admin/index.ts
@@ -61,3 +61,12 @@ export type {
   TaskModalState,
   WeekModalState,
 } from './useProgramRoadmap';
+
+export { useInvites, isRowValid, EMAIL_REGEX, VALID_ROLES } from './useInvites';
+export type {
+  InviteStatusFilter,
+  InviteRecord,
+  CreatedInvite,
+  CsvRow,
+  BulkReport,
+} from './useInvites';
